style: re-apply analyzer/style cleanup over merged incoming code
Three-phase dotnet format (whitespace/style/analyzers, info severity)
constrained to the 206 files introduced by the fork+master merge
(redundancy/HA/CTT/complex-types). 87 files normalized.

Reverted 4 files where dotnet format broke compilation (documented hazards):
- ClientReplicaCoordinatorTests.cs (merge-marker injection, CS8300)
- ManagedSessionTests.cs (async stripped from await method, CS4032/CS0161)
- SharedKeyValueMonitoredItemQueueFactoryTests.cs / SharedKeyValueSubscriptionStoreTests.cs
  (async stripped from async-iterator, CS8403/CS8424 on net472/net48)
Manually wrapped one RCS0056 over-long Assert in KubernetesLeaseLeaderElectionTests.cs.

+++ Applications/RedundantPubSub/Program.cs
@@ -38,7 +38,6 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Opc.Ua;
 using Opc.Ua.PubSub.Configuration;
-using Opc.Ua.PubSub.Encoding;
 using Opc.Ua.PubSub.Redundancy;
 using Opc.Ua.Redundancy;
 using Raft;
@@ -179,7 +178,7 @@ namespace RedundantPubSub
             EmitSimulatedSequences(
                 ownerId: "publisher-b",
                 start: demoOptions.HaMode == PubSubRedundancyMode.Hot
-                    ? (uint)simulatedPreFailoverCount + simulatedPostFailoverMargin
+                    ? simulatedPreFailoverCount + simulatedPostFailoverMargin
                     : 0,
                 count: simulatedPostFailoverCount,
                 ref simulatedLastSequence,
